Move GetColumnIndex method
